misc/HtmlTestSuite/results/test-0532.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="https://github.com/w3c/csswg-drafts/issues/7660" rel="help"/>
<style>
.target {
  display: grid;
  grid-auto-flow: column;
  position: relative;
  line-height: 0;
  font-size: 20px;
  inline-size: 300px;
  margin-block: 10px;
  padding: 10px;
  border: solid 3px;
  writing-mode: vertical-rl;
}

.inner {
  overflow: hidden;
  inline-size: 100px;
  block-size: 80px;
  margin: 10px;
  border: solid 5px;
  padding: 10px;
  font-size: 30px;
}

span {
  display: inline-block;
  width: 1em;
  height: 1em;
  outline: solid cyan 3px;
  outline-offset: -3px;
}
</style>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
</head><body onload="checkLayout('.target &gt; *')">

<div class="target" style="align-items: first baseline;">
  <div data-offset-x="70"><span/><br/><span/></div>
  <div class="inner" data-offset-x="20">
    <div>
      <span/><br/><span/>
    </div>
  </div>
</div>

<div class="target" style="align-items: last baseline;">
  <div data-offset-x="60"><span/><br/><span/></div>
  <div class="inner" data-offset-x="20">
    <div>
      <span/><br/><span/>
    </div>
  </div>
</div>

<div class="target" style="align-items: first baseline;">
  <div data-offset-x="100"><span/><br/><span/></div>
  <div class="inner" data-offset-x="20">
    <div style="margin-block-start: -200px;">
      <span/><br/><span/>
    </div>
  </div>
</div>

<div class="target" style="align-items: last baseline;">
  <div data-offset-x="120"><span/><br/><span/></div>
  <div class="inner" data-offset-x="20">
    <div style="margin-block-start: -200px;">
      <span/><br/><span/>
    </div>
  </div>
</div>

<div class="target" style="align-items: first baseline;">
  <div data-offset-x="10"><span/><br/><span/></div>
  <div class="inner" data-offset-x="40">
    <div style="margin-block-start: 200px;">
      <span/><br/><span/>
    </div>
  </div>
</div>

<div class="target" style="align-items: last baseline;">
  <div data-offset-x="10"><span/><br/><span/></div>
  <div class="inner" data-offset-x="20">
    <div style="margin-block-start: 200px;">
      <span/><br/><span/>
    </div>
  </div>
</div>
</body></html>
